pyproject.toml
[tool.poetry]
name = "course-work-6"
version = "0.1.0"
description = ""
authors = ["vitt <fivtin@mail.ru>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
django = "4.2.2"
psycopg2-binary = "^2.9.9"
python-dotenv = "^1.0.1"
redis = "^5.0.6"
django-apscheduler = "^0.6.2"
pytils = "^0.4.1"
pillow = "^10.3.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
